refactor: merge state manager and snapshot
