+++ src/control_plane/container/reconciler.ts
@@ -1,20 +1,24 @@
 import { Config } from '#self/config';
 import { Clock } from '#self/lib/clock';
 import { DependencyContext } from '#self/lib/dependency_context';
+import { EventBus } from '#self/lib/event-bus';
 import { Logger, loggers } from '#self/lib/loggers';
 import { createDeferred } from '#self/lib/util';
+import { ContainerReconciledEvent } from '../events';
 import { ContainerManager } from './container_manager';
 
 export type ReconcilerContext = {
   config: Config;
   clock: Clock;
   containerManager: ContainerManager;
+  eventBus: EventBus;
 };
 
 export class ContainerReconciler {
   private _reconcilingInterval: number;
   private _clock: Clock;
   private _containerManager: ContainerManager;
+  private _eventBus: EventBus;
   private _logger: Logger;
   private _interval: unknown | null = null;
   private _closed = false;
@@ -26,6 +30,7 @@ export class ContainerReconciler {
     this._reconcilingInterval =
       ctx.getInstance('config').turf.reconcilingInterval;
     this._containerManager = ctx.getInstance('containerManager');
+    this._eventBus = ctx.getInstance('eventBus');
     this._logger = loggers.get('container reconciler');
   }
 
@@ -63,6 +68,9 @@ export class ContainerReconciler {
     this._clock.clearTimeout(this._interval);
     this._containerManager
       .reconcileContainers()
+      .then(() => {
+        return this._eventBus.publish(new ContainerReconciledEvent());
+      })
       .catch(err => {
         this._logger.error('unexpected error on reconciliation', err);
       })
