--- src/control_plane/worker_stats/index.ts
@@ -1,3 +1,0 @@
-export * from './snapshot';
-export * from './broker';
-export * from './worker';
