--- src/control_plane/worker_stats/snapshot.ts
@@ -1,205 +1,0 @@
-import EventEmitter from 'events';
-import { BaseOf } from '#self/lib/sdk_base';
-import { Broker } from './broker';
-import loggers from '#self/lib/logger';
-import { Logger } from '#self/lib/loggers';
-import { Worker, WorkerMetadata } from './worker';
-import { FunctionProfileManager } from '#self/lib/function_profile';
-import { Config } from '#self/config';
-import * as root from '#self/proto/root';
-import { WorkerStatus } from '#self/lib/constants';
-import { StatLogger } from './stat_logger';
-
-export class WorkerStatsSnapshot extends BaseOf(EventEmitter) {
-  private logger: Logger;
-  private profiles: FunctionProfileManager;
-  public brokers: Map<string, Broker>;
-  private statLogger: StatLogger;
-
-  constructor(profileManager: FunctionProfileManager, public config: Config) {
-    super();
-
-    this.logger = loggers.get('worker_stats snapshot');
-    this.profiles = profileManager;
-
-    this.statLogger = new StatLogger();
-
-    this.brokers = new Map();
-  }
-
-  /**
-   * Sync data from data plane.
-   */
-  sync(syncData: root.noslated.data.IBrokerStats[]) {
-    const newMap: Map<string, Broker> = new Map();
-    for (const item of syncData) {
-      const key = Broker.getKey(item.functionName!, item.inspector!);
-      const broker = this.getBroker(item.functionName!, item.inspector!);
-      if (!broker) {
-        // 一切以 Control Plane 已存在数据为准
-        continue;
-      }
-
-      broker.sync(item.workers!);
-      newMap.set(key, broker);
-      this.brokers.delete(key);
-    }
-
-    for (const [key, value] of this.brokers.entries()) {
-      value.sync([]);
-      newMap.set(key, value);
-    }
-
-    this.brokers = newMap;
-  }
-
-  /**
-   * Get broker by function name and `isInspector`.
-   */
-  getBroker(functionName: string, isInspector: boolean): Broker | null {
-    return this.brokers.get(Broker.getKey(functionName, isInspector)) || null;
-  }
-
-  /**
-   * Get or create broker by function name and `isInspector`.
-   */
-  getOrCreateBroker(
-    functionName: string,
-    isInspector: boolean,
-    disposable = false
-  ): Broker | null {
-    let broker = this.getBroker(functionName, isInspector);
-    if (broker) return broker;
-    if (!this.profiles.get(functionName)) return null;
-    broker = new Broker(
-      this.profiles,
-      this.config,
-      functionName,
-      isInspector,
-      disposable
-    );
-    this.brokers.set(Broker.getKey(functionName, isInspector), broker);
-    return broker;
-  }
-
-  /**
-   * Get worker object.
-   */
-  getWorker(
-    functionName: string,
-    isInspector: boolean,
-    workerName: string
-  ): Worker | null {
-    const broker = this.getBroker(functionName, isInspector);
-    if (!broker) return null;
-    return broker.getWorker(workerName) || null;
-  }
-
-  /**
-   * Register worker.
-   * @param {string} funcName The function name.
-   * @param {string} processName The process name (worker name).
-   * @param {string} credential The credential.
-   * @param {boolean} isInspector Whether it's using inspector or not.
-   */
-  register(workerMetadata: WorkerMetadata): Worker {
-    const broker = this.getOrCreateBroker(
-      workerMetadata.funcName,
-      workerMetadata.options.inspect,
-      workerMetadata.disposable
-    );
-    if (!broker) {
-      throw new Error(
-        `No function named ${workerMetadata.funcName} in function profile.`
-      );
-    }
-    return broker.register(workerMetadata);
-  }
-
-  toProtobufObject(): root.noslated.data.IBrokerStats[] {
-    return [...this.brokers.values()].map(broker => broker.toJSON());
-  }
-
-  /**
-   * TODO: resource manager
-   * Try worker GC.
-   * @param {Broker} broker The broker object.
-   * @param {Worker} worker The worker object.
-   * @return {Promise<void>} The result.
-   */
-  async #tryGC(broker: Broker, worker: Worker) {
-    if (!broker.getWorker(worker.name)) {
-      throw new Error(
-        `Worker ${worker.name} not belongs to broker ${Broker.getKey(
-          broker.name,
-          broker.isInspector
-        )}`
-      );
-    }
-
-    // 进入该状态，必然要被 GC
-    if (
-      worker.workerStatus === WorkerStatus.Stopped ||
-      worker.workerStatus === WorkerStatus.Unknown
-    ) {
-      try {
-        await worker.container!.stop();
-      } catch (e) {
-        this.logger.warn(
-          `Failed to stop worker [${worker.name}] via \`.#tryGC()\`.`,
-          e
-        );
-      }
-
-      const state = await worker.container!.terminated;
-      if (state) {
-        const stime = state['rusage.stime'] ?? 0;
-        const utime = state['rusage.utime'] ?? 0;
-        //TODO(yilong.lyl): fix typo @zl131478
-        const rss = state['rusage.masrss'];
-
-        this.statLogger.exit(
-          state.name,
-          state.pid,
-          state.exitcode ?? null,
-          state['killed.signal'] ?? null,
-          stime + utime,
-          rss,
-          worker.requestId ?? null
-        );
-      }
-
-      this.logger.info("%s's last state: %j", worker.name, state);
-      broker.removeItemFromStartingPool(worker.name);
-      this.emit('workerStopped', state, broker, worker);
-
-      broker.workers.delete(worker.name);
-    }
-  }
-
-  /**
-   * Correct synced data (remove GCed items)
-   * @return {Promise<void>} The result.
-   */
-  async correct() {
-    const gcs = [];
-    for (const broker of this.brokers.values()) {
-      for (const worker of broker.workers.values()) {
-        gcs.push(this.#tryGC(broker, worker));
-      }
-    }
-
-    const result = await Promise.allSettled(gcs);
-    for (const it of result) {
-      if (it.status === 'rejected') {
-        this.logger.error('unexpected error on gc', it.reason);
-      }
-    }
-
-    for (const broker of [...this.brokers.values()]) {
-      if (!broker.workers.size && !broker.data) {
-        this.brokers.delete(Broker.getKey(broker.name, broker.isInspector));
-      }
-    }
-  }
-}
